[chore] App.js에서 프로필 context 삭제 및  나만의 코스 Context 분리
