+++ OMO-project/src/App.jsx
@@ -33,31 +33,9 @@ import MyCourseDetail from "./pages/MyCourse/MyCourseDetail/MyCourseDetail";
 import {useEffect, useState} from "react";
 import {dataCopy} from "./const/dataCopy";
 import LoginLoading from "./pages/LoginLoading/LoginLoading";
-import { ProfileImageProvider } from "./assets/context/ProfileImageContext";
+import {MyCourseProvider} from "./assets/context/MyCourseContext";
 
-const reducer = (state, action) => {
-  let newState = [];
-  switch (action.type) {
-    case "INIT": {
-      return action.data;
-    }
-    case "CREATE": {
-      newState = [action.data, ...state];
-      break;
-    }
-    default:
-      return state;
-  }
-  localStorage.setItem("mycourseboard", JSON.stringify(newState));
-  return newState;
-};
-
-export const MyCourseStateContext = React.createContext();
-export const MyCourseDispatchContext = React.createContext();
-
-const App = ({ handleLogout, isLoggedIn, setIsLoggedIn }) => {
-
-
+const App = ({handleLogout, isLoggedIn, setIsLoggedIn}) => {
   // 현재 위치를 가져오기 위한 구글 API KEY (메인페이지에 사용 - 검색창에 디폴트로 현재 위치 뜨게)
   const GOOGLE_MAPS_API_KEY = "AIzaSyBFZH53aP29Zr7vY5jyv7wd4wGQMg3CI1s";
 
@@ -97,139 +75,93 @@ const App = ({ handleLogout, isLoggedIn, setIsLoggedIn }) => {
       });
   }, []);
 
-  // 나만의 코스 data
-  const [data, dispatch] = useReducer(reducer, []);
-
-  // 나만의 코스 list
-  useEffect(() => {
-    const localData = localStorage.getItem("mycourseboard");
-    if (localData) {
-      const boardList = JSON.parse(localData).sort((a, b) => parseInt(b.reg_at) - parseInt(a.reg_at));
-      if (boardList.length >= 1) {
-        dataId.current = parseInt(boardList[0].id) + 1;
-        dispatch({type: "INIT", data: boardList});
-      }
-    }
-  }, []);
-
   // 마이페이지 최근 본 장소 상태 관리
   const [recentData, setRecentData] = useState([]);
-
-  const dataId = useRef(0);
-
-  // 나만의 코스 CREATE
-  const onCreate = (dates, title, content, regat) => {
-    dispatch({
-      type: "CREATE",
-      data: {
-        id: dataId.current,
-        title,
-        dates,
-        content,
-        regat,
-      },
-    });
-    dataId.current += 1;
-  };
 
   const [defaultListImg, setDefaultListImg] = useState("/src/assets/detail/defaultDetailIcon.png");
 
   return (
-    <MyCourseStateContext.Provider value={data}>
-      <MyCourseDispatchContext.Provider
-        value={{
-          onCreate,
-        }}
-      >
- <ProfileImageProvider>
-          <div>
-            {/* 헤더 */}
-            <Header isLoggedIn={isLoggedIn} handleLogout={handleLogout}/>
-            <Routes>
-              {/* 메인 페이지 */}
-              <Route path="/" element={<Main setSearchResultsX={setSearchResultsX} setSearchResultsY={setSearchResultsY} location={location} setLocation={setLocation} />} />
+    <MyCourseProvider>
+      <div>
+        {/* 헤더 */}
+        <Header isLoggedIn={isLoggedIn} handleLogout={handleLogout} />
+        <Routes>
+          {/* 메인 페이지 */}
+          <Route path="/" element={<Main setSearchResultsX={setSearchResultsX} setSearchResultsY={setSearchResultsY} location={location} setLocation={setLocation} />} />
 
-              {/* 서브 페이지 */}
-              <Route path="/Eating" element={<Eating />} />
-              <Route path="/Watching" element={<Watching />} />
-              <Route path="/Playing" element={<Playing />} />
-              <Route path="/ThemeCafe" element={<ThemeCafe />} />
+          {/* 서브 페이지 */}
+          <Route path="/Eating" element={<Eating />} />
+          <Route path="/Watching" element={<Watching />} />
+          <Route path="/Playing" element={<Playing />} />
+          <Route path="/ThemeCafe" element={<ThemeCafe />} />
 
-              {/* 로그인/회원가입 */}
-              <Route path="/Login" element={<Login />} />
-              <Route path="/Signup" element={<Signup />} />
-              <Route path="/LoginLoading" element={<LoginLoading />} />
+          {/* 로그인/회원가입 */}
+          <Route path="/Login" element={<Login />} />
+          <Route path="/Signup" element={<Signup />} />
+          <Route path="/LoginLoading" element={<LoginLoading />} />
 
-              {/* 리스트페이지 */}
-              <Route
-                path="/List"
-                element={
-                  <List
-                    recentData={recentData}
-                    setRecentData={setRecentData}
-                    dataCopy={dataCopy}
-                    searchResultsX={searchResultsX}
-                    searchResultsY={searchResultsY}
-                    defaultListImg={defaultListImg}
-                    setDefaultListImg={setDefaultListImg}
-                  />
-                }
+          {/* 리스트페이지 */}
+          <Route
+            path="/List"
+            element={
+              <List
+                recentData={recentData}
+                setRecentData={setRecentData}
+                dataCopy={dataCopy}
+                searchResultsX={searchResultsX}
+                searchResultsY={searchResultsY}
+                defaultListImg={defaultListImg}
+                setDefaultListImg={setDefaultListImg}
               />
-              <Route
-                path="/List/:category"
-                element={
-                  <List
-                    recentData={recentData}
-                    setRecentData={setRecentData}
-                    dataCopy={dataCopy}
-                    searchResultsX={searchResultsX}
-                    searchResultsY={searchResultsY}
-                    defaultListImg={defaultListImg}
-                    setDefaultListImg={setDefaultListImg}
-                  />
-                }
+            }
+          />
+          <Route
+            path="/List/:category"
+            element={
+              <List
+                recentData={recentData}
+                setRecentData={setRecentData}
+                dataCopy={dataCopy}
+                searchResultsX={searchResultsX}
+                searchResultsY={searchResultsY}
+                defaultListImg={defaultListImg}
+                setDefaultListImg={setDefaultListImg}
               />
+            }
+          />
 
-              {/* 상세페이지 */}
-              <Route
-                path="/DetailMenu/:id/:place_name"
-                element={
-                  <DetailMenu defaultListImg={defaultListImg} setDefaultListImg={setDefaultListImg} />
-                }
-              />
-              <Route path="/DetailNone" element={<DetailNone />} />
-              <Route path="/DetailTariff" element={<DetailTariff />} />
+          {/* 상세페이지 */}
+          <Route path="/DetailMenu/:id/:place_name" element={<DetailMenu defaultListImg={defaultListImg} setDefaultListImg={setDefaultListImg} />} />
+          <Route path="/DetailNone" element={<DetailNone />} />
+          <Route path="/DetailTariff" element={<DetailTariff />} />
 
-              {/* 마이 페이지 */}
-             
-              <Route path="/MyInfo" element={<MyInfo />} />
-              <Route path="/Interest" element={<Interest />} />
-              <Route path="/Recommend" element={<Recommend />} />
-              <Route path="/Recent" element={<Recent recentData={recentData} />} />
-              <Route path="/MyWrote" element={<MyWrote />} />
-              <Route path="/ProfileSetting" element={<ProfileSetting isLoggedIn={isLoggedIn} setIsLoggedIn={setIsLoggedIn}/>} />
-             
+          {/* 마이 페이지 */}
 
-              {/* 나만의 코스 */}
-              <Route path="/MyCourseMain" element={<MyCourseMain />} />
-              <Route path="/MyCourseNewWrite" element={<MyCourseNewWrite />} />
-              <Route path="/MyCourseDetail/:id" element={<MyCourseDetail />} />
-              <Route path="/MyCourseOthersVersion/:id" element={<MyCourseOthersVersion />} />
+          <Route path="/MyInfo" element={<MyInfo />} />
+          <Route path="/Interest" element={<Interest />} />
+          <Route path="/Recommend" element={<Recommend />} />
+          <Route path="/Recent" element={<Recent recentData={recentData} />} />
+          <Route path="/MyWrote" element={<MyWrote />} />
+          <Route path="/ProfileSetting" element={<ProfileSetting isLoggedIn={isLoggedIn} setIsLoggedIn={setIsLoggedIn} />} />
 
-              {/* 커뮤니티 */}
-              <Route path="/MyCourseBoard" element={<MyCourseBoard />} />
-              <Route path="/WorryBoard" element={<WorryBoard />} />
-              <Route path="/FreeBoard" element={<FreeBoard />} />
-              <Route path="/InquiryBoardFrequent" element={<InquiryBoardFrequent />} />
-              <Route path="/InquiryBoardQnA" element={<InquiryBoardQnA />} />
+          {/* 나만의 코스 */}
+          <Route path="/MyCourseMain" element={<MyCourseMain />} />
+          <Route path="/MyCourseNewWrite" element={<MyCourseNewWrite />} />
+          <Route path="/MyCourseDetail/:id" element={<MyCourseDetail />} />
+          <Route path="/MyCourseOthersVersion/:id" element={<MyCourseOthersVersion />} />
 
-              {/* 공지사항 */}
-              <Route path="/Notice" element={<Notice />} />
-            </Routes>
-          </div>
-          </ProfileImageProvider>
-      </MyCourseDispatchContext.Provider>
-    </MyCourseStateContext.Provider>
+          {/* 커뮤니티 */}
+          <Route path="/MyCourseBoard" element={<MyCourseBoard />} />
+          <Route path="/WorryBoard" element={<WorryBoard />} />
+          <Route path="/FreeBoard" element={<FreeBoard />} />
+          <Route path="/InquiryBoardFrequent" element={<InquiryBoardFrequent />} />
+          <Route path="/InquiryBoardQnA" element={<InquiryBoardQnA />} />
+
+          {/* 공지사항 */}
+          <Route path="/Notice" element={<Notice />} />
+        </Routes>
+      </div>
+    </MyCourseProvider>
   );
 };
 export default App;
